Crate signer

This crate serves as interface between native frontend and Rust code. Try to avoid placing any logic here, just interfacing. When porting to new platform, all Rust changes will probably happen here.

Macros

uniffi_reexport_scaffolding

Enums

ErrorDisplayed

Container for severe error message

QrSequenceDecodeError

An error type for QR sequence decoding errors.

Functions

ffi_signer_12c4_rustbuffer_alloc

ffi_signer_12c4_rustbuffer_free⚠

ffi_signer_12c4_rustbuffer_from_bytes⚠

ffi_signer_12c4_rustbuffer_reserve⚠

signer_destroy_string⚠

Public destructor for strings managed by the other side of the FFI.