Crate signer

This crate serves as interface between native frontend and Rust code. Try to avoid placing any logic here, just interfacing. When porting to new platform, all Rust changes will probably happen here.

Macros

uniffi_reexport_scaffolding

Enums

ErrorDisplayed

Container for severe error message

Functions

ffi_signer_e31b_rustbuffer_alloc

ffi_signer_e31b_rustbuffer_free⚠

ffi_signer_e31b_rustbuffer_from_bytes⚠

ffi_signer_e31b_rustbuffer_reserve⚠

signer_destroy_string⚠

Public destructor for strings managed by the other side of the FFI.